Fold common files into platform specific packages
Since the content of opx-platform-config package is so small (it is
just a few hundred bytes) it doesn't make much sense to maintain it
as a separate package.

Another significant advantage of this change is that those files can
be replaced/overriden if needed by new target platforms. This avoids
impacting packaging for unrelated platforms.